requirements_rock_ultralytics.txt
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.1
fonttools==4.53.0
fsspec==2024.6.0
idna==3.7
importlib-resources==6.4.0
jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-periphery==2.4.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.12.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.32
ultralytics-thop==0.2.8
urllib3==2.2.1
zipp==3.19.2
